Move common platform code up one level
